reorganized data extraction logic into separate module and classes
